+++ app/src/organisms/ODD/QuickTransferFlow/Aspirate/constants.ts
@@ -1,0 +1,11 @@
+export const SETTING_OPTIONS = {
+  ASPIRATE_FLOW_RATE: 'aspirate_flow_rate',
+  ASPIRATE_TIP_POSITION: 'aspirate_tip_position',
+  ASPIRATE_SUBMERGE: 'aspirate_submerge',
+  PRE_WET_TIP: 'pre_wet_tip',
+  ASPIRATE_MIX: 'aspirate_mix',
+  ASPIRATE_DELAY: 'aspirate_delay',
+  ASPIRATE_RETRACT: 'aspirate_retract',
+  ASPIRATE_TOUCH_TIP: 'aspirate_touch_tip',
+  ASPIRATE_AIR_GAP: 'aspirate_air_gap',
+} as const
