+++ app/src/organisms/ODD/QuickTransferFlow/Aspirate/index.tsx
@@ -15,23 +15,9 @@ import type {
   QuickTransferSummaryAction,
   QuickTransferSummaryState,
 } from '../types'
+import type { SettingOption } from './types'
 
 const PADDING_TOP_FOR_NAV = '12rem'
-
-export const SETTING_OPTIONS = {
-  ASPIRATE_FLOW_RATE: 'aspirate_flow_rate',
-  ASPIRATE_TIP_POSITION: 'aspirate_tip_position',
-  ASPIRATE_SUBMERGE: 'aspirate_submerge',
-  PRE_WET_TIP: 'pre_wet_tip',
-  ASPIRATE_MIX: 'aspirate_mix',
-  ASPIRATE_DELAY: 'aspirate_delay',
-  ASPIRATE_RETRACT: 'aspirate_retract',
-  ASPIRATE_TOUCH_TIP: 'aspirate_touch_tip',
-  ASPIRATE_AIR_GAP: 'aspirate_air_gap',
-} as const
-
-export type SettingOption = typeof SETTING_OPTIONS[keyof typeof SETTING_OPTIONS]
-
 interface AspirateProps {
   state: QuickTransferSummaryState
   dispatch: Dispatch<QuickTransferSummaryAction>
