+++ app/src/organisms/ODD/QuickTransferFlow/Aspirate/types.ts
@@ -1,0 +1,2 @@
+import type { SETTING_OPTIONS } from './constants'
+export type SettingOption = typeof SETTING_OPTIONS[keyof typeof SETTING_OPTIONS]
